requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-core==2.10.1
python-multipart==0.0.6
boto3==1.28.85
google-cloud-speech==2.24.1
ffmpeg-python==0.2.0
static-ffmpeg
google-cloud-storage==2.10.0
aiohttp>=3.8.0
